Refactor Article entity to use decimal for Stock and MinimumStock properties
- Updated ArticleDto and Article entity to change Stock and MinimumStock types from int to decimal for improved precision in inventory management.
- Modified ArticleConfiguration to set appropriate database precision for Stock and MinimumStock fields.
- Ensured consistency across DTO, entity, and database configuration for article management.

+++ backend/src/Spisa.Application/DTOs/ArticleDto.cs
@@ -8,8 +8,8 @@ public class ArticleDto
     public long CategoryId { get; set; }
     public string CategoryName { get; set; } = string.Empty;
     public decimal UnitPrice { get; set; }
-    public int Stock { get; set; }
-    public int MinimumStock { get; set; }
+    public decimal Stock { get; set; }
+    public decimal MinimumStock { get; set; }
     public string? Location { get; set; }
     public bool IsDiscontinued { get; set; }
     public string? Notes { get; set; }
